pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-guard-kit"
version = "0.17.0"
description = "v0.17.0: MiniJudge ($0 AUROC 0.747, distilled from Sonnet), cross-domain TV validated AUROC 0.660 [CI 0.614-0.705] n=1000, QuickCalibrator, 4 platform integrations (Langfuse/LangSmith/Prometheus/Datadog). Real-time reliability monitoring for LLM agents. v0.16.1: fix FastAPI>=0.115 (starlette 0.52 compat), domain-invariant feature selection 0.778 cross-domain. v0.16.0: hosted calibration endpoint, multilevel features, 3-domain validation, latency SLA. v0.15.0: probe_ensemble_blend() +1.6pp AUROC."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Avighan Majumder"},
]
keywords = ["llm", "reliability", "failure-prediction", "anomaly-detection", "knn",
            "claude", "openai", "nvidia", "nim", "langchain", "llamaindex", "haystack", "monitoring"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "numpy>=1.20",
    "scikit-learn>=0.24",
    "sentence-transformers>=2.2.0",
    "anthropic>=0.7.0",
    "openai>=1.0",
]

[project.optional-dependencies]
langchain  = ["langchain-core>=0.1"]
nim        = ["openai>=1.0"]
llamaindex = ["llama-index-core>=0.10"]
haystack   = ["haystack-ai>=2.0"]
qara       = ["torch>=1.13"]
server     = ["fastapi>=0.115", "uvicorn[standard]>=0.23", "pydantic>=2.0"]
saas       = ["stripe>=8.0", "httpx>=0.27", "pyjwt>=2.8"]
mcp        = ["mcp>=1.0"]
dev        = ["pytest>=7.0", "matplotlib>=3.4", "httpx>=0.24"]
all = [
    "langchain-core>=0.1",
    "openai>=1.0",
    "llama-index-core>=0.10",
    "haystack-ai>=2.0",
    "torch>=1.13",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.23",
    "pydantic>=2.0",
    "passlib[bcrypt]>=1.7",
    "stripe>=8.0",
    "httpx>=0.27",
    "pyjwt>=2.8",
    "mcp>=1.0",
]

[project.scripts]
"llm-guard-kit" = "qppg_service.cli:main"
"llm-guard-mcp"  = "llm_guard_mcp.server:run"

[project.urls]
Repository = "https://github.com/avighan/qppg"
Issues = "https://github.com/avighan/qppg/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["qppg*", "llm_guard*", "qppg_service*", "llm_guard_mcp*"]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m not slow')",
    "integration: marks tests requiring live model or network access",
]
